refactor(compare): share normalization helpers
Move compare normalization and canonicalization into a shared library module so direct compare and compare-outputs use the same alert-reduction logic.
